moved routing to separate file
