+++ server/index.js
@@ -1,7 +1,7 @@
 const app = require('./app.js');
 
 const PORT = 3001;
-const HOST = 'localhost';
+const HOST = '0.0.0.0';
 
 app.listen(PORT, HOST, () => {
   console.log(`listening at http://${HOST}:${PORT}`)
